+++ moodle-plugin/local/minerva/db/events.php
@@ -26,14 +26,6 @@ defined('MOODLE_INTERNAL') || die();
 
 $observers = [
     [
-        'eventname' => '\core\event\user_enrolment_created',
-        'callback' => '\local_minerva\observer::user_enrolled',
-    ],
-    [
-        'eventname' => '\core\event\user_enrolment_deleted',
-        'callback' => '\local_minerva\observer::user_unenrolled',
-    ],
-    [
         'eventname' => '\core\event\course_deleted',
         'callback' => '\local_minerva\observer::course_deleted',
     ],